function-download-url/requirements.txt
google-cloud-storage
functions-framework
Flask